pyproject.toml
[tool.poetry]
name = "statistic-api"
version = "0.1.0"
description = "statistic-api"
authors = ["Hide <padocon@naver.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11.7"
alembic = "^1.13.1"
uvicorn = "^0.25.0"
fastapi = "^0.109.0"
celery = "^5.3.6"
gunicorn = "^21.2.0"
fastapi-event = "^0.1.3"
pythondi = "^1.2.4"
ujson = "^5.9.0"
aiomysql = "^0.2.0"
click = "^8.1.7"
redis = "^5.0.1"
pydantic-settings = "^2.1.0"
pyjwt = "^2.8.0"
cryptography = "^41.0.7"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
pymysql = "^1.1.0"
dependency-injector = "^4.41.0"
httpx = "^0.26.0"
anyio = "^3.6.1"
ciso8601 = "^2.2.0"
h11 = "^0.13.0"
idna = "^3.3"
loguru = "^0.6.0"
pydantic = "^2.7.0"
python-dotenv = "^0.21.0"
pytz = "^2022.1"
tzlocal = "^4.2"
psycopg2-binary = "^2.9.9" 
asyncpg = "^0.29.0"
clickhouse-cityhash = "^1.0.2.3"
clickhouse-sqlalchemy = "^0.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
types-redis = "^4.6.0.20240106"
coverage = "^7.4.0"

[tool.coverage.run]
omit = ["tests/*", "test_*.py", "migrations/*"]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
